Modularize basic-page content-type component
