--- src/components/BasicPage.vue
@@ -1,26 +1,0 @@
-<template>
-
-</template>
-
-<script lang="ts">
-import {Component, Prop, Vue} from 'vue-property-decorator';
-
-@Component({
-    components: {
-
-    },
-})
-
-// Description of Vue component
-export default class ComponentTitle extends Vue {
-
-    constructor() {
-        super();
-    }
-
-}
-</script>
-
-<style lang="scss" scoped>
-
-</style>
